+++ embedchain/loaders/index.ts
@@ -1,6 +1,6 @@
-import { LocalQnaPairLoader } from "./local_qna_pair";
-import { PdfFileLoader } from "./pdf_file";
-import { WebPageLoader } from "./web_page";
+import { LocalQnaPairLoader } from "./LocalQnaPair";
+import { PdfFileLoader } from "./PdfFile";
+import { WebPageLoader } from "./WebPage";
 import { BaseLoader } from "./BaseLoader";
 
 export {LocalQnaPairLoader, PdfFileLoader, WebPageLoader, BaseLoader};
